+++ src/lib/jukette.ts
@@ -26,6 +26,7 @@ export type {
 	JuketteEventDetail,
 	JuketteEventName,
 	JuketteMidiOscillator,
+	JuketteSoundCloudPreload,
 	JuketteTrack,
 	JuketteTrackKind,
 	MidiNote,